Assessment Task partially done - contains error

+++ database/migrations/2022_05_13_220658_create_affiliates_table.php
@@ -18,7 +18,8 @@ return new class extends Migration
             $table->foreignId('user_id');
             $table->foreignId('merchant_id');
             // TODO: Replace me with a brief explanation of why floats aren't the correct data type, and replace with the correct data type.
-            $table->float('commission_rate');
+            // In PHP; By default floating values have data types of double so arithmetic operations with decimal numbers are automatically handled as doubles, making them more convenient to use in many cases.
+            $table->double('commission_rate');
             $table->string('discount_code');
             $table->timestamps();
         });
